#27 done 3 btn đầu, còn đọc tiếp khi nào xong lịch sử sẽ làm
