+++ myapp/templates/theloai.html
@@ -18,7 +18,7 @@
 								<img src="{{x.anhbia.url}}" alt="{{x.ten}}" loading="lazy" width="320" height="480">
 							</a>
 						</div>
-						<div class="inner-name">
+						<div class=" inner-name">
 							<a href="/truyen_id={{x.id}}/">
 								<div class="description">{{x.ten}} - Ch.{{x.chapmoinhat.stt}}</div>
 							</a>
